requirements.txt
streamlit==1.28.1
requests==2.31.0
anthropic==0.7.8
python-dotenv==1.0.0